refactor: organize dist_alerts specific analytics code
